fix migration tests
