+++ unit_tests/sources/declarative/spec/test_spec.py
@@ -2,8 +2,6 @@
 # Copyright (c) 2025 Airbyte, Inc., all rights reserved.
 #
 
-from unittest.mock import Mock, mock_open
-
 import pytest
 
 from airbyte_cdk.models import (
@@ -159,108 +157,6 @@ from airbyte_cdk.sources.declarative.validators.validate_adheres_to_schema impor
 )
 def test_spec(spec, expected_connection_specification) -> None:
     assert spec.generate_spec() == expected_connection_specification
-
-
-@pytest.fixture
-def migration_mocks(monkeypatch):
-    mock_message_repository = Mock()
-    mock_message_repository.consume_queue.return_value = [Mock()]
-
-    mock_source = Mock()
-    mock_entrypoint = Mock()
-    mock_entrypoint.extract_config.return_value = "/fake/config/path"
-    monkeypatch.setattr(
-        "airbyte_cdk.sources.declarative.spec.spec.AirbyteEntrypoint", lambda _: mock_entrypoint
-    )
-
-    _mock_open = mock_open()
-    mock_json_dump = Mock()
-    mock_print = Mock()
-    mock_serializer_dump = Mock()
-
-    mock_decoded_bytes = Mock()
-    mock_decoded_bytes.decode.return_value = "decoded_message"
-    mock_orjson_dumps = Mock(return_value=mock_decoded_bytes)
-
-    monkeypatch.setattr("builtins.open", _mock_open)
-    monkeypatch.setattr("json.dump", mock_json_dump)
-    monkeypatch.setattr("builtins.print", mock_print)
-    monkeypatch.setattr(
-        "airbyte_cdk.models.airbyte_protocol_serializers.AirbyteMessageSerializer.dump",
-        mock_serializer_dump,
-    )
-    monkeypatch.setattr("airbyte_cdk.sources.declarative.spec.spec.orjson.dumps", mock_orjson_dumps)
-
-    return {
-        "message_repository": mock_message_repository,
-        "source": mock_source,
-        "open": _mock_open,
-        "json_dump": mock_json_dump,
-        "print": mock_print,
-        "serializer_dump": mock_serializer_dump,
-        "orjson_dumps": mock_orjson_dumps,
-        "decoded_bytes": mock_decoded_bytes,
-    }
-
-
-def test_given_unmigrated_config_when_migrating_then_config_is_migrated(migration_mocks) -> None:
-    input_config = {"planet": "CRSC"}
-
-    spec = component_spec(
-        connection_specification={},
-        parameters={},
-        config_migrations=[
-            ConfigMigration(
-                description="Test migration",
-                transformations=[
-                    ConfigRemapField(
-                        map={"CRSC": "Coruscant"}, field_path=["planet"], config=input_config
-                    )
-                ],
-            )
-        ],
-    )
-    spec.message_repository = migration_mocks["message_repository"]
-
-    spec.migrate_config("/fake/config/path", input_config)
-
-    migration_mocks["message_repository"].emit_message.assert_called_once()
-    migration_mocks["open"].assert_called_once_with("/fake/config/path", "w")
-    migration_mocks["json_dump"].assert_called_once()
-    migration_mocks["print"].assert_called()
-    migration_mocks["serializer_dump"].assert_called()
-    migration_mocks["orjson_dumps"].assert_called()
-    migration_mocks["decoded_bytes"].decode.assert_called()
-
-
-def test_given_already_migrated_config_no_control_message_is_emitted(migration_mocks) -> None:
-    input_config = {"planet": "Coruscant"}
-
-    spec = component_spec(
-        connection_specification={},
-        parameters={},
-        config_migrations=[
-            ConfigMigration(
-                description="Test migration",
-                transformations=[
-                    ConfigRemapField(
-                        map={"CRSC": "Coruscant"}, field_path=["planet"], config=input_config
-                    )
-                ],
-            )
-        ],
-    )
-    spec.message_repository = migration_mocks["message_repository"]
-
-    spec.migrate_config("/fake/config/path", input_config)
-
-    migration_mocks["message_repository"].emit_message.assert_not_called()
-    migration_mocks["open"].assert_not_called()
-    migration_mocks["json_dump"].assert_not_called()
-    migration_mocks["print"].assert_not_called()
-    migration_mocks["serializer_dump"].assert_not_called()
-    migration_mocks["orjson_dumps"].assert_not_called()
-    migration_mocks["decoded_bytes"].decode.assert_not_called()
 
 
 def test_given_list_of_transformations_when_transform_config_then_config_is_transformed() -> None:
