+++ unit_tests/sources/declarative/test_manifest_declarative_source.py
@@ -9,7 +9,7 @@ import sys
 from copy import deepcopy
 from pathlib import Path
 from typing import Any, List, Mapping
-from unittest.mock import call, patch
+from unittest.mock import Mock, call, mock_open, patch
 
 import pytest
 import requests
@@ -30,6 +30,9 @@ from airbyte_cdk.models import (
 )
 from airbyte_cdk.sources.declarative.declarative_stream import DeclarativeStream
 from airbyte_cdk.sources.declarative.manifest_declarative_source import ManifestDeclarativeSource
+from airbyte_cdk.sources.declarative.parsers.model_to_component_factory import (
+    ModelToComponentFactory,
+)
 from airbyte_cdk.sources.declarative.retrievers.simple_retriever import SimpleRetriever
 
 logger = logging.getLogger("airbyte")
@@ -2099,3 +2102,180 @@ def test_slice_checkpoint(test_name, manifest, pages, expected_states_qty):
     with patch.object(SimpleRetriever, "_fetch_next_page", side_effect=pages):
         states = [message.state for message in _run_read(manifest, _stream_name) if message.state]
         assert len(states) == expected_states_qty
+
+
+@pytest.fixture
+def migration_mocks(monkeypatch):
+    mock_message_repository = Mock()
+    mock_message_repository.consume_queue.return_value = [Mock()]
+
+    _mock_open = mock_open()
+    mock_json_dump = Mock()
+    mock_print = Mock()
+    mock_serializer_dump = Mock()
+
+    mock_decoded_bytes = Mock()
+    mock_decoded_bytes.decode.return_value = "decoded_message"
+    mock_orjson_dumps = Mock(return_value=mock_decoded_bytes)
+
+    monkeypatch.setattr("builtins.open", _mock_open)
+    monkeypatch.setattr("json.dump", mock_json_dump)
+    monkeypatch.setattr("builtins.print", mock_print)
+    monkeypatch.setattr(
+        "airbyte_cdk.models.airbyte_protocol_serializers.AirbyteMessageSerializer.dump",
+        mock_serializer_dump,
+    )
+    monkeypatch.setattr(
+        "airbyte_cdk.sources.declarative.manifest_declarative_source.orjson.dumps",
+        mock_orjson_dumps,
+    )
+
+    return {
+        "message_repository": mock_message_repository,
+        "open": _mock_open,
+        "json_dump": mock_json_dump,
+        "print": mock_print,
+        "serializer_dump": mock_serializer_dump,
+        "orjson_dumps": mock_orjson_dumps,
+        "decoded_bytes": mock_decoded_bytes,
+    }
+
+
+def test_given_unmigrated_config_when_migrating_then_config_is_migrated(migration_mocks) -> None:
+    input_config = {"planet": "CRSC"}
+
+    manifest = {
+        "version": "0.34.2",
+        "type": "DeclarativeSource",
+        "check": {"type": "CheckStream", "stream_names": ["Test"]},
+        "streams": [
+            {
+                "type": "DeclarativeStream",
+                "name": "Test",
+                "schema_loader": {
+                    "type": "InlineSchemaLoader",
+                    "schema": {"type": "object"},
+                },
+                "retriever": {
+                    "type": "SimpleRetriever",
+                    "requester": {
+                        "type": "HttpRequester",
+                        "url_base": "https://example.org",
+                        "path": "/test",
+                        "authenticator": {"type": "NoAuth"},
+                    },
+                    "record_selector": {
+                        "type": "RecordSelector",
+                        "extractor": {"type": "DpathExtractor", "field_path": []},
+                    },
+                },
+            }
+        ],
+        "spec": {
+            "type": "Spec",
+            "documentation_url": "https://example.org",
+            "connection_specification": {},
+            "config_normalization_rules": {
+                "config_migrations": [
+                    {
+                        "type": "ConfigMigration",
+                        "description": "Test migration",
+                        "transformations": [
+                            {
+                                "type": "ConfigRemapField",
+                                "map": {"CRSC": "Coruscant"},
+                                "field_path": ["planet"],
+                            }
+                        ],
+                    }
+                ],
+            },
+        },
+    }
+
+    ManifestDeclarativeSource(
+        source_config=manifest,
+        config=input_config,
+        config_path="/fake/config/path",
+        component_factory=ModelToComponentFactory(
+            message_repository=migration_mocks["message_repository"],
+        ),
+    )
+
+    migration_mocks["message_repository"].emit_message.assert_called_once()
+    migration_mocks["open"].assert_called_once_with("/fake/config/path", "w")
+    migration_mocks["json_dump"].assert_called_once()
+    migration_mocks["print"].assert_called()
+    migration_mocks["serializer_dump"].assert_called()
+    migration_mocks["orjson_dumps"].assert_called()
+    migration_mocks["decoded_bytes"].decode.assert_called()
+
+
+def test_given_already_migrated_config_no_control_message_is_emitted(migration_mocks) -> None:
+    input_config = {"planet": "Coruscant"}
+
+    manifest = {
+        "version": "0.34.2",
+        "type": "DeclarativeSource",
+        "check": {"type": "CheckStream", "stream_names": ["Test"]},
+        "streams": [
+            {
+                "type": "DeclarativeStream",
+                "name": "Test",
+                "schema_loader": {
+                    "type": "InlineSchemaLoader",
+                    "schema": {"type": "object"},
+                },
+                "retriever": {
+                    "type": "SimpleRetriever",
+                    "requester": {
+                        "type": "HttpRequester",
+                        "url_base": "https://example.org",
+                        "path": "/test",
+                        "authenticator": {"type": "NoAuth"},
+                    },
+                    "record_selector": {
+                        "type": "RecordSelector",
+                        "extractor": {"type": "DpathExtractor", "field_path": []},
+                    },
+                },
+            }
+        ],
+        "spec": {
+            "type": "Spec",
+            "documentation_url": "https://example.org",
+            "connection_specification": {},
+            "config_normalization_rules": {
+                "config_migrations": [
+                    {
+                        "type": "ConfigMigration",
+                        "description": "Test migration",
+                        "transformations": [
+                            {
+                                "type": "ConfigRemapField",
+                                "map": {"CRSC": "Coruscant"},
+                                "field_path": ["planet"],
+                            }
+                        ],
+                    }
+                ],
+            },
+        },
+    }
+
+    ManifestDeclarativeSource(
+        source_config=manifest,
+        config=input_config,
+        config_path="/fake/config/path",
+        component_factory=ModelToComponentFactory(
+            message_repository=migration_mocks["message_repository"],
+        ),
+    )
+
+    migration_mocks["message_repository"].emit_message.assert_not_called()
+    migration_mocks["open"].assert_not_called()
+    migration_mocks["json_dump"].assert_not_called()
+    migration_mocks["print"].assert_not_called()
+    migration_mocks["serializer_dump"].assert_not_called()
+    migration_mocks["orjson_dumps"].assert_not_called()
+    migration_mocks["decoded_bytes"].decode.assert_not_called()
